Разграничил классы, пересчитал координаты дверей
