+++ .gitignore
@@ -6,3 +6,6 @@ venv/
 /Lib/
 /Traffic-Control-project-3/
 /Test input video/
+/Test video output/
+/self development images/
+/self development dataset/
